+++ src/index.js
@@ -9,6 +9,8 @@ module.exports = {
   PermissionLevels: require('./structures/PermissionLevels'),
   Pages: require('./structures/Pages'),
   MySql: require('./structures/MySql'),
+  Store: require('./structures/Store'),
   Monitor: require('./structures/Monitor'),
   Trigger: require('./structures/Trigger'),
+  Inhibitor: require('./structures/Inhibitor'),
 };