+++ src/structures/Trigger.js
@@ -2,6 +2,9 @@ const Base = require('./Base');
 
 const defaultOptions = {};
 
+/**
+ * @extends {Base}
+ */
 module.exports = class Trigger extends Base {
   constructor(client, store, fullpath, options = {}) {
     super(client, store, 'trigger', fullpath, options);